working on fieldSize sync

+++ nestjs/conf/src/chat/service/connected-user/connected-user.service.ts
@@ -37,11 +37,13 @@ export class ConnectedUserService {
   }
 
   async deleteBySocketId(socketId: string): Promise<ConnectedUser> {
-    return this.prisma.connectedUser.delete({
-      where: {
-        socketId,
-      },
-    });
+	if (this.prisma.connectedUser){
+		return this.prisma.connectedUser.delete({
+		  where: {
+			socketId,
+		  },
+		});
+	}
   }
 
   async deleteAll() {
